Code refactoring and highlight selected tab (#5)
* feat: separate task manager into components

* feat: restructure component hierarchy

* fix: uninstall zustand (will be part of future PR)

* fix: revert package json changes
