--- src/components/TaskManager.tsx
@@ -1,87 +1,0 @@
-import React, { useState } from "react";
-
-import TaskItem from "./TaskItem";
-import { Task } from "../types/task";
-
-type Filter = "all" | "completed" | "pending";
-
-const TaskManager = () => {
-  const [tasks, setTasks] = useState<Task[]>([
-    { id: 1, title: "Buy groceries", completed: false },
-    { id: 2, title: "Clean the house", completed: true },
-  ]);
-  const [filter, setFilter] = useState<Filter>("all");
-  const [newTaskTitle, setNewTaskTitle] = useState<string>("");
-
-  const filteredTasks = tasks.filter((task) => {
-    if (filter === "completed") return task.completed === true;
-    if (filter === "pending") return task.completed === false;
-    return true;
-  });
-
-  const handleAddTask = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (newTaskTitle.trim() === "") return;
-    const newTask: Task = {
-      id: tasks.length + 1,
-      title: newTaskTitle,
-      completed: false,
-    };
-    setTasks([...tasks, newTask]);
-    setNewTaskTitle("");
-  };
-
-  const handleDeleteTask = (id: number) => {
-    setTasks(tasks.filter((task) => task.id !== id));
-  };
-
-  const toggleTaskCompletion = (id: number) => {
-    const updatedTasks = tasks.map((task) =>
-      task.id === id ? { ...task, completed: !task.completed } : task
-    );
-    setTasks(updatedTasks);
-  };
-
-  return (
-    <div className="container mx-auto bg-white p-4 rounded shadow">
-      <form onSubmit={handleAddTask} className="mb-4 flex">
-        <input
-          type="text"
-          placeholder="New task..."
-          value={newTaskTitle}
-          onChange={(e) => setNewTaskTitle(e.target.value)}
-          className="flex-grow border rounded-l py-2 px-3"
-        />
-        <button type="submit" className="bg-blue-500 text-white px-4 rounded-r">
-          Add
-        </button>
-      </form>
-      <div className="flex justify-around mb-4">
-        <button onClick={() => setFilter("all")} className="text-gray-700">
-          All
-        </button>
-        <button
-          onClick={() => setFilter("completed")}
-          className="text-gray-700"
-        >
-          Completed
-        </button>
-        <button onClick={() => setFilter("pending")} className="text-gray-700">
-          Pending
-        </button>
-      </div>
-      <ul>
-        {filteredTasks.map((task) => (
-          <TaskItem
-            key={task.id}
-            task={task}
-            onDelete={handleDeleteTask}
-            onToggle={toggleTaskCompletion}
-          />
-        ))}
-      </ul>
-    </div>
-  );
-};
-
-export default TaskManager;
