+++ src/types/filter.ts
@@ -1,0 +1,1 @@
+export type Filter = "all" | "completed" | "pending";
